+++ src/KOKKOS/npair_kokkos.cpp
@@ -425,9 +425,9 @@ void NeighborKokkosExecute<DeviceType>::
 
   const AtomNeighbors neighbors_i = neigh_transpose ?
     neigh_list.get_neighbors_transpose(i) : neigh_list.get_neighbors(i);
-  const double xtmp = x(i, 0);
-  const double ytmp = x(i, 1);
-  const double ztmp = x(i, 2);
+  const double xtmp = static_cast<double>(x(i, 0));
+  const double ytmp = static_cast<double>(x(i, 1));
+  const double ztmp = static_cast<double>(x(i, 2));
   const int itype = type(i);
   tagint itag;
   if (HalfNeigh && Newton && Tri) itag = tag(i);
@@ -447,22 +447,22 @@ void NeighborKokkosExecute<DeviceType>::
 
     if (j <= i) continue;
     if (j >= nlocal) {
-      if (x(j,2) < ztmp) continue;
-      if (x(j,2) == ztmp) {
-        if (x(j,1) < ytmp) continue;
-        if (x(j,1) == ytmp && x(j,0) < xtmp) continue;
+      if (static_cast<double>(x(j,2)) < ztmp) continue;
+      if (static_cast<double>(x(j,2)) == ztmp) {
+        if (static_cast<double>(x(j,1)) < ytmp) continue;
+        if (static_cast<double>(x(j,1)) == ytmp && static_cast<double>(x(j,0)) < xtmp) continue;
       }
     }
 
     const int jtype = type(j);
     if (exclude && exclusion(i,j,itype,jtype)) continue;
 
-    const double delx = xtmp - x(j, 0);
-    const double dely = ytmp - x(j, 1);
-    const double delz = ztmp - x(j, 2);
+    const double delx = xtmp - static_cast<double>(x(j, 0));
+    const double dely = ytmp - static_cast<double>(x(j, 1));
+    const double delz = ztmp - static_cast<double>(x(j, 2));
     const double rsq = delx*delx + dely*dely + delz*delz;
 
-    if (rsq <= cutneighsq(itype,jtype)) {
+    if (rsq <= static_cast<double>(cutneighsq(itype,jtype))) {
       if (molecular != Atom::ATOMIC) {
         if (!moltemplate)
           which = find_special(i,j);
@@ -517,12 +517,12 @@ void NeighborKokkosExecute<DeviceType>::
             } else if (itag < jtag) {
               if ((itag+jtag) % 2 == 1) continue;
             } else {
-              if (fabs(x(j,2)-ztmp) > delta) {
-                if (x(j,2) < ztmp) continue;
-              } else if (fabs(x(j,1)-ytmp) > delta) {
-                if (x(j,1) < ytmp) continue;
+              if (fabs(static_cast<double>(x(j,2))-ztmp) > delta) {
+                if (static_cast<double>(x(j,2)) < ztmp) continue;
+              } else if (fabs(static_cast<double>(x(j,1))-ytmp) > delta) {
+                if (static_cast<double>(x(j,1)) < ytmp) continue;
               } else {
-                if (x(j,0) < xtmp) continue;
+                if (static_cast<double>(x(j,0)) < xtmp) continue;
               }
             }
           }
@@ -531,12 +531,12 @@ void NeighborKokkosExecute<DeviceType>::
         const int jtype = type(j);
         if (exclude && exclusion(i,j,itype,jtype)) continue;
 
-        const double delx = xtmp - x(j, 0);
-        const double dely = ytmp - x(j, 1);
-        const double delz = ztmp - x(j, 2);
+        const double delx = xtmp - static_cast<double>(x(j, 0));
+        const double dely = ytmp - static_cast<double>(x(j, 1));
+        const double delz = ztmp - static_cast<double>(x(j, 2));
         const double rsq = delx*delx + dely*dely + delz*delz;
 
-        if (rsq <= cutneighsq(itype,jtype)) {
+        if (rsq <= static_cast<double>(cutneighsq(itype,jtype))) {
           if (molecular != Atom::ATOMIC) {
             if (!moltemplate)
               which = NeighborKokkosExecute<DeviceType>::find_special(i,j);
@@ -672,10 +672,10 @@ void NeighborKokkosExecute<DeviceType>::build_ItemGPU(typename Kokkos::TeamPolic
 
         if (j <= i) continue;
         if (j >= nlocal) {
-          if (x(j,2) < ztmp) continue;
-          if (x(j,2) == ztmp) {
-            if (x(j,1) < ytmp) continue;
-            if (x(j,1) == ytmp && x(j,0) < xtmp) continue;
+          if (static_cast<double>(x(j,2)) < ztmp) continue;
+          if (static_cast<double>(x(j,2)) == ztmp) {
+            if (static_cast<double>(x(j,1)) < ytmp) continue;
+            if (static_cast<double>(x(j,1)) == ytmp && static_cast<double>(x(j,0)) < xtmp) continue;
           }
         }
 
@@ -687,7 +687,7 @@ void NeighborKokkosExecute<DeviceType>::build_ItemGPU(typename Kokkos::TeamPolic
         const double delz = ztmp - other_x[m + 2 * atoms_per_bin];
         const double rsq = delx*delx + dely*dely + delz*delz;
 
-        if (rsq <= cutneighsq(itype,jtype)) {
+        if (rsq <= static_cast<double>(cutneighsq(itype,jtype))) {
           if (molecular != Atom::ATOMIC) {
             int which = 0;
             if (!moltemplate)
@@ -764,12 +764,12 @@ void NeighborKokkosExecute<DeviceType>::build_ItemGPU(typename Kokkos::TeamPolic
               } else if (itag < jtag) {
                 if ((itag+jtag) % 2 == 1) continue;
               } else {
-                if (fabs(x(j,2)-ztmp) > delta) {
-                  if (x(j,2) < ztmp) continue;
-                } else if (fabs(x(j,1)-ytmp) > delta) {
-                  if (x(j,1) < ytmp) continue;
+                if (fabs(static_cast<double>(x(j,2))-ztmp) > delta) {
+                  if (static_cast<double>(x(j,2)) < ztmp) continue;
+                } else if (fabs(static_cast<double>(x(j,1))-ytmp) > delta) {
+                  if (static_cast<double>(x(j,1)) < ytmp) continue;
                 } else {
-                  if (x(j,0) < xtmp) continue;
+                  if (static_cast<double>(x(j,0)) < xtmp) continue;
                 }
               }
             }
@@ -783,7 +783,7 @@ void NeighborKokkosExecute<DeviceType>::build_ItemGPU(typename Kokkos::TeamPolic
           const double delz = ztmp - other_x[m + 2 * atoms_per_bin];
           const double rsq = delx*delx + dely*dely + delz*delz;
 
-          if (rsq <= cutneighsq(itype,jtype)) {
+          if (rsq <= static_cast<double>(cutneighsq(itype,jtype))) {
             if (molecular != Atom::ATOMIC) {
               int which = 0;
               if (!moltemplate)
@@ -846,9 +846,9 @@ void NeighborKokkosExecute<DeviceType>::
 
   const AtomNeighbors neighbors_i = neigh_transpose ?
     neigh_list.get_neighbors_transpose(i) : neigh_list.get_neighbors(i);
-  const double xtmp = x(i, 0);
-  const double ytmp = x(i, 1);
-  const double ztmp = x(i, 2);
+  const double xtmp = static_cast<double>(x(i, 0));
+  const double ytmp = static_cast<double>(x(i, 1));
+  const double ztmp = static_cast<double>(x(i, 2));
   const int itype = type(i);
 
   const typename AT::t_int_1d_const_um stencil
@@ -874,12 +874,12 @@ void NeighborKokkosExecute<DeviceType>::
         const int jtype = type[j];
         if (exclude && exclusion(i,j,itype,jtype)) continue;
 
-        const double delx = xtmp - x(j,0);
-        const double dely = ytmp - x(j,1);
-        const double delz = ztmp - x(j,2);
+        const double delx = xtmp - static_cast<double>(x(j,0));
+        const double dely = ytmp - static_cast<double>(x(j,1));
+        const double delz = ztmp - static_cast<double>(x(j,2));
         const double rsq = delx*delx + dely*dely + delz*delz;
 
-        if (rsq <= cutneighsq(itype,jtype)) {
+        if (rsq <= static_cast<double>(cutneighsq(itype,jtype))) {
           if (molecular != Atom::ATOMIC) {
             if (!moltemplate)
               which = find_special(i,j);
@@ -929,12 +929,12 @@ void NeighborKokkosExecute<DeviceType>::
         const int jtype = type[j];
         if (exclude && exclusion(i,j,itype,jtype)) continue;
 
-        const double delx = xtmp - x(j,0);
-        const double dely = ytmp - x(j,1);
-        const double delz = ztmp - x(j,2);
+        const double delx = xtmp - static_cast<double>(x(j,0));
+        const double dely = ytmp - static_cast<double>(x(j,1));
+        const double delz = ztmp - static_cast<double>(x(j,2));
         const double rsq = delx*delx + dely*dely + delz*delz;
 
-        if (rsq <= cutneighsq(itype,jtype)) {
+        if (rsq <= static_cast<double>(cutneighsq(itype,jtype))) {
           if (n < neigh_list.maxneighs) neighbors_i(n++) = j;
           else n++;
         }
@@ -1078,7 +1078,7 @@ void NeighborKokkosExecute<DeviceType>::build_ItemGhostGPU(typename Kokkos::Team
           const double delz = ztmp - other_x[m + 2 * atoms_per_bin];
           const double rsq = delx*delx + dely*dely + delz*delz;
 
-          if (rsq <= cutneighsq(itype,jtype)) {
+          if (rsq <= static_cast<double>(cutneighsq(itype,jtype))) {
             if (molecular != Atom::ATOMIC && !ghost) {
               if (!moltemplate)
                 which = NeighborKokkosExecute<DeviceType>::find_special(i,j);
@@ -1136,10 +1136,10 @@ void NeighborKokkosExecute<DeviceType>::
 
   const AtomNeighbors neighbors_i = neigh_transpose ?
     neigh_list.get_neighbors_transpose(i) : neigh_list.get_neighbors(i);
-  const double xtmp = x(i, 0);
-  const double ytmp = x(i, 1);
-  const double ztmp = x(i, 2);
-  const double radi = radius(i);
+  const double xtmp = static_cast<double>(x(i, 0));
+  const double ytmp = static_cast<double>(x(i, 1));
+  const double ztmp = static_cast<double>(x(i, 2));
+  const double radi = static_cast<double>(radius(i));
   const int itype = type(i);
   tagint itag;
   if (HalfNeigh && Newton && Tri) itag = tag(i);
@@ -1162,21 +1162,21 @@ void NeighborKokkosExecute<DeviceType>::
 
     if (j <= i) continue;
     if (j >= nlocal) {
-      if (x(j,2) < ztmp) continue;
-      if (x(j,2) == ztmp) {
-        if (x(j,1) < ytmp) continue;
-        if (x(j,1) == ytmp && x(j,0) < xtmp) continue;
+      if (static_cast<double>(x(j,2)) < ztmp) continue;
+      if (static_cast<double>(x(j,2)) == ztmp) {
+        if (static_cast<double>(x(j,1)) < ytmp) continue;
+        if (static_cast<double>(x(j,1)) == ytmp && static_cast<double>(x(j,0)) < xtmp) continue;
       }
     }
 
     const int jtype = type(j);
     if (exclude && exclusion(i,j,itype,jtype)) continue;
 
-    const double delx = xtmp - x(j, 0);
-    const double dely = ytmp - x(j, 1);
-    const double delz = ztmp - x(j, 2);
+    const double delx = xtmp - static_cast<double>(x(j, 0));
+    const double dely = ytmp - static_cast<double>(x(j, 1));
+    const double delz = ztmp - static_cast<double>(x(j, 2));
     const double rsq = delx*delx + dely*dely + delz*delz;
-    const double radsum = radi + radius(j);
+    const double radsum = radi + static_cast<double>(radius(j));
     const double cutsq = (radsum + skin) * (radsum + skin);
 
     if (rsq <= cutsq) {
@@ -1242,12 +1242,12 @@ void NeighborKokkosExecute<DeviceType>::
           } else if (itag < jtag) {
             if ((itag+jtag) % 2 == 1) continue;
           } else {
-            if (fabs(x(j,2)-ztmp) > delta) {
-              if (x(j,2) < ztmp) continue;
-            } else if (fabs(x(j,1)-ytmp) > delta) {
-              if (x(j,1) < ytmp) continue;
+            if (fabs(static_cast<double>(x(j,2))-ztmp) > delta) {
+              if (static_cast<double>(x(j,2)) < ztmp) continue;
+            } else if (fabs(static_cast<double>(x(j,1))-ytmp) > delta) {
+              if (static_cast<double>(x(j,1)) < ytmp) continue;
             } else {
-              if (x(j,0) < xtmp) continue;
+              if (static_cast<double>(x(j,0)) < xtmp) continue;
             }
           }
         }
@@ -1256,11 +1256,11 @@ void NeighborKokkosExecute<DeviceType>::
       const int jtype = type(j);
       if (exclude && exclusion(i,j,itype,jtype)) continue;
 
-      const double delx = xtmp - x(j, 0);
-      const double dely = ytmp - x(j, 1);
-      const double delz = ztmp - x(j, 2);
+      const double delx = xtmp - static_cast<double>(x(j, 0));
+      const double dely = ytmp - static_cast<double>(x(j, 1));
+      const double delz = ztmp - static_cast<double>(x(j, 2));
       const double rsq = delx*delx + dely*dely + delz*delz;
-      const double radsum = radi + radius(j);
+      const double radsum = radi + static_cast<double>(radius(j));
       const double cutsq = (radsum + skin) * (radsum + skin);
 
       if (rsq <= cutsq) {
@@ -1390,10 +1390,10 @@ void NeighborKokkosExecute<DeviceType>::build_ItemSizeGPU(typename Kokkos::TeamP
 
         if (j <= i) continue;
         if (j >= nlocal) {
-          if (x(j,2) < ztmp) continue;
-          if (x(j,2) == ztmp) {
-            if (x(j,1) < ytmp) continue;
-            if (x(j,1) == ytmp && x(j,0) < xtmp) continue;
+          if (static_cast<double>(x(j,2)) < ztmp) continue;
+          if (static_cast<double>(x(j,2)) == ztmp) {
+            if (static_cast<double>(x(j,1)) < ytmp) continue;
+            if (static_cast<double>(x(j,1)) == ytmp && static_cast<double>(x(j,0)) < xtmp) continue;
           }
         }
 
@@ -1491,12 +1491,12 @@ void NeighborKokkosExecute<DeviceType>::build_ItemSizeGPU(typename Kokkos::TeamP
               } else if (itag < jtag) {
                 if ((itag+jtag) % 2 == 1) continue;
               } else {
-                if (fabs(x(j,2)-ztmp) > delta) {
-                  if (x(j,2) < ztmp) continue;
-                } else if (fabs(x(j,1)-ytmp) > delta) {
-                  if (x(j,1) < ytmp) continue;
+                if (fabs(static_cast<double>(x(j,2))-ztmp) > delta) {
+                  if (static_cast<double>(x(j,2)) < ztmp) continue;
+                } else if (fabs(static_cast<double>(x(j,1))-ytmp) > delta) {
+                  if (static_cast<double>(x(j,1)) < ytmp) continue;
                 } else {
-                  if (x(j,0) < xtmp) continue;
+                  if (static_cast<double>(x(j,0)) < xtmp) continue;
                 }
               }
             }
